pyproject.toml
[project]
name = "openmeals"
version = "1.0.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    # web
    "fastapi[standard]~=0.135.0",
    "uvicorn[standard]~=0.42.0",      # pulls uvloop + httptools automatically
    # db
    "sqlmodel~=0.0.37",
    "sqlalchemy[asyncio]~=2.0.48",
    "asyncpg~=0.31.0",
    "psycopg[binary]~=3.2.0",
    "alembic~=1.18.4",
    # config/validation
    "pydantic~=2.12.5",
    "pydantic-settings~=2.13.1",
    # auth
    "authlib~=1.6.9",
    "httpx",
    # ulid
    "python-ulid~=3.1.0",
    # vector search
    "pgvector~=0.4.0",
    # task queue
    "celery[redis]~=5.4.0",
    # ml-depedencies. in default for simplicity, but must be used only in worker scope.
    "torch",
    "sentence-transformers~=4.0.0",
    # observability
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-instrumentation-celery",
    "opentelemetry-instrumentation-httpx",
    "boto3~=1.42.73",
    "pillow~=12.1.1",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "ipython",
    "polyfactory",
    "pytest-cov>=7.1.0",
    "boto3-stubs[s3]>=1.42.73",
]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "FAST", "S", "C4","T20", "PTH", "I", "N", "PERF", "PLE", "PLC",]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if t.TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
omit = [
    "src/core/celery.py",
    "src/core/lifespan.py",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S"]
"**/routers/**/*.py" = ["FAST003"]
"**/handlers.py" = ["FAST003"]
